Initial commit
This repo contains information/code that is the result of my
research/experiments with the 433MHz protocol used by Kidde KN-COSM-B-RF
smoke/CO alarms and the Wink Hub.

None of the information/code in this repository is suitable for
real-world use in life safety applications. It is for
educational/informational purposes only.