Add firmware for ATmega328P (WIP)
This is a more polished firmware based on the STM32duino sketch. It was
developed against a modified RM1101-USB-232 dongle with the ATmega48PA
swapped out for an ATmega328P. Jumpers were added to route hardware SPI
on the ATmega to the CC1101. See https://imgur.com/a/ixWAovH.

Example serial output (JSON):

{"millis":54761,"type":"alarm","command":"smoke","min_rssi":25,"max_rssi":25,"duration":0,"count":1,"active":true}
{"millis":65441,"type":"alarm","command":"smoke","min_rssi":25,"max_rssi":25,"duration":5552,"count":3,"active":false}

Note that this is not considered complete or even tested at this point.

DISCLAIMER: This is not intended to be used for any real-world safety
applications - it is for educational/informational purposes only! Use at
your own risk.